Provide custom URL for integration tests
This change makes it possible to specify a new environment variable,
MEDIAWIKI_API_URL, that will be used as the endpoint for the
integration tests. This can then be set in a local phpunit.xml file,
for situations in which a local development installation of MediaWiki
is installed somewhere other than at http://localhost/w/

Also, some phpdocs have been aded to the TestEnvironment class,
and some extra files to be ignored by git.