Continue all PageListGetter operations
This change makes all methods of the PageListGetter class
continue their queries (where the first request doesn't
retrieve the whole result set). It also adds tests for this
class, and refactors it to move the actual request-loop
code into a new protected method.

In addition, a bug [1] in the getRandom() method is fixed.

[1] https://phabricator.wikimedia.org/T142459