Run integration tests on Travis and fix failures
This adds a system of running integration tests on Travis, and
fixes the following bugs that were preventing the tests from
passing:

The continue parameter in PageListGetter::getFromWhatLinksHere()
was not configured as a generator, and didn't have a extraParams
parameter.

The UserCreator was not working with recent MediaWiki.

Also added better page-creation and job running logic (although,
the latter could certainly be improved) to the CategoryTraverserTest.

Bug: https://phabricator.wikimedia.org/T111690